refactor: new commit_verify crate started from scratch
