+++ commit_verify.old/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "commit_verify"
+version = "0.9.0"
+license = "Apache-2.0"
+authors = ["Dr. Maxim Orlovsky <orlovsky@lnp-bp.org>"]
+description = "Commit-verify API for client-side validation"
+repository = "https://github.com/LNP-BP/client_side_validation"
+homepage = "https://github.com/LNP-BP"
+keywords = ["lnp-bp", "cryptography", "smart-contracts", "commit-verify"]
+categories = ["cryptography", "encoding"]
+readme = "README.md"
+edition = "2021"
+rust-version = "1.59.0"
+
+[lib]
+name = "commit_verify"
+path = "src/lib.rs"
+
+[dependencies]
+amplify = { version = "4.0.0-beta.2", features = ["hex"] }
+bitcoin_hashes = "0.11.0"
+strict_encoding = "2.0.0-beta.2"
+rand = { version = "0.8.5", optional = true }
+serde_crate = { version = "1.0", package = "serde", optional = true }
+serde_with = { version = "1.14.0", optional = true }
+lnpbp_secp256k1zkp = { version = "0.9.0", optional = true }
+
+[dev-dependencies]
+rand = "0.8.5"
+
+[features]
+default = []
+all = ["rand", "serde", "bulletproofs"]
+serde = ["serde_crate", "serde_with", "bitcoin_hashes/serde", "amplify/serde"]
+bulletproofs = ["lnpbp_secp256k1zkp"]
+
+[package.metadata.docs.rs]
+features = [ "all" ]
