+++ commit_verify.old/src/commit_encode.rs
